+++ src/weather/mod.rs
@@ -48,6 +48,7 @@ pub trait LookUpCity {
 mod open_meteo;
 mod open_weather_map;
 mod provider;
+mod tomorrow_io;
 pub mod tools;
 mod weather_api;
 mod weather_bit;